New ui elements (plots, freq table)

+++ views/expoview.hpp
@@ -15,13 +15,13 @@ protected:
     Methods createMethods();
 signals:
     void type(QString status, QString type);
-    void run(QString status, unsigned int nRun);
+    void run(QString status, unsigned nRun);
     void launchRead(QString status);
     void stopRead(QString status);
     void launchFreq(QString status);
     void stopFreq(QString status);
-    void triggerCount(QString status, unsigned count, unsigned drop);
-    void packageCount(QString status, unsigned count, unsigned drop);
+    void triggerCount(QString status, uintmax_t count, uintmax_t drop);
+    void packageCount(QString status, uintmax_t count, uintmax_t drop);
     void chambersCount(QString status, TrekFreq count, TrekFreq drop);
     void freq(QString status, TrekFreq freq);
 };
